requirement.txt
flask
psycopg2-binary
python-dotenv
Flask-JWT-Extended
bcrypt